requirements.txt
pdfplumber==0.10.3
Pillow==10.1.0
sentence-transformers==3.4.1
chromadb==0.6.3
openai==1.30.1
numpy==1.26.2
streamlit==1.33.0
pandas==2.1.1
python-dotenv==1.0.1
PyMuPDF==1.23.3
plotly==5.18.0